Add a [(rust.closed_enum)=true] option.
This option replaces the normally generated open enum with its `_Closed`
variant. That means the enum is required to be a known variant at
deserialization time. This makes using enums more ergonomic for
situations where version skew is not relevant, or is dealt with some
other way (e.g. migrations)