Rewrite codegen in Rust (#164)
This completely removes the dependency on Python, and especially the Python protobuf package.

pb-jelly-gen now depends on code that it generated for the various descriptors used by protoc as well as `rust/extensions.proto`. These are bundled into the crate using a newly added "single file" output mode to avoid needing to publish separate crates. The generated code is also checked in for bootstrapping.

For now this is mostly a 1:1 straight port of the original Python code. Boxing recursive fields was changed to use a side set as it's hard to mutate the original protos in-place (and pb-jelly support for extensions is read-only anyway).

Using pb-jelly-gen as a library no longer needs to create virtualenvs or run the plugin separately; instead it uses protoc's support for dumping parsed protos into a FileDescriptorSet and invokes the codegen "plugin" directly. A binary plugin target is still provided for direct use with protoc though.